requirements.txt
tensorflow-gpu==2.0.0-rc0
gast==0.2.2
